forgotpwd-works with errors
