--- src/screens/PasswordResetScreen.js
@@ -1,108 +1,0 @@
-import React, { useState } from 'react';
-import { View, Text, TextInput, TouchableOpacity, StyleSheet } from 'react-native';
-import { Feather } from '@expo/vector-icons';
-
-const PasswordResetScreen = () => {
-  const [newPassword, setNewPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const [showNewPassword, setShowNewPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-
-  const handleProceed = () => {
-    // Implement password reset logic
-    if (newPassword === confirmPassword) {
-      console.log('Password reset successful');
-    } else {
-      console.log('Passwords do not match');
-    }
-  };
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Password Recovery</Text>
-      <Text style={styles.subtitle}>
-        Create a new password to access your account
-      </Text>
-
-      <View style={styles.passwordContainer}>
-        <TextInput
-          style={styles.passwordInput}
-          placeholder="New Password"
-          value={newPassword}
-          onChangeText={setNewPassword}
-          secureTextEntry={!showNewPassword}
-        />
-        <TouchableOpacity onPress={() => setShowNewPassword(!showNewPassword)} style={styles.eyeIcon}>
-          <Feather name={showNewPassword ? 'eye' : 'eye-off'} size={24} color="gray" />
-        </TouchableOpacity>
-      </View>
-
-      <View style={styles.passwordContainer}>
-        <TextInput
-          style={styles.passwordInput}
-          placeholder="Confirm Password"
-          value={confirmPassword}
-          onChangeText={setConfirmPassword}
-          secureTextEntry={!showConfirmPassword}
-        />
-        <TouchableOpacity onPress={() => setShowConfirmPassword(!showConfirmPassword)} style={styles.eyeIcon}>
-          <Feather name={showConfirmPassword ? 'eye' : 'eye-off'} size={24} color="gray" />
-        </TouchableOpacity>
-      </View>
-
-      <TouchableOpacity style={styles.proceedButton} onPress={handleProceed}>
-        <Text style={styles.proceedButtonText}>Proceed</Text>
-      </TouchableOpacity>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: 'white',
-    marginTop: 40.58,
-
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: 21,
-  },
-  subtitle: {
-    fontSize: 13,
-    color: 'black',
-    textAlign: 'center',
-    marginBottom: 20,
-  },
-  passwordContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderColor: '#ddd',
-    borderRadius: 8,
-    marginBottom: 20,
-  },
-  passwordInput: {
-    flex: 1,
-    padding: 10,
-  },
-  eyeIcon: {
-    padding: 10,
-  },
-  proceedButton: {
-    backgroundColor: 'forestgreen',
-    padding: 15,
-    borderRadius: 71,
-    alignItems: 'center',
-    marginTop: 30,
-  },
-  proceedButtonText: {
-    color: 'white',
-    fontWeight: 'bold',
-  },
-});
-
-export default PasswordResetScreen;
