+++ src/screens/checkEmail.js
@@ -1,0 +1,113 @@
+import React, { useState } from 'react';
+import { Text, TextInput, TouchableOpacity, StyleSheet, SafeAreaView, Alert, ActivityIndicator } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
+import axios from 'axios';
+
+import API_URL from '../config/api'
+
+
+const checkEmail = () => {
+  const [email, setEmail] = useState('');
+  const [loading, setLoading] = useState(false);
+  const navigation = useNavigation();
+
+  
+  const handleProceed = async () => {
+    if (!email.trim()) {
+      Alert.alert('Error', 'Please enter a valid email address.');
+      return;
+    }
+
+    setLoading(true);
+    try {
+      const response = await axios.post(`${API_URL}/emailcheck`, { email }, {
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+
+      console.log(response.data);
+
+      if (response.status === 200) {
+        Alert.alert('Success', 'Email exists! Proceeding to reset password.');
+        navigation.navigate('PasswordReset', { email });
+      } else if (response.status === 404) {
+        Alert.alert('Error', 'No account found with this email address.');
+      } else {
+        Alert.alert('Error', 'Something went wrong. Please try again.');
+      }
+    } catch (error) {
+      console.error(error);
+      Alert.alert('Error', 'Unable to connect to the server. Please try again.');
+    } finally {
+      setLoading(false);
+    }
+  };
+
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <Text style={styles.title}>Password Recovery</Text>
+      <Text style={styles.subtitle}>
+        Enter your email address to recover your password
+      </Text>
+
+      <TextInput
+        style={styles.input}
+        placeholder="Email"
+        value={email}
+        onChangeText={setEmail}
+        keyboardType="email-address"
+        autoCapitalize="none"
+      />
+
+      <TouchableOpacity style={styles.proceedButton} onPress={handleProceed} disabled={loading}>
+        {loading ? (
+          <ActivityIndicator color="white" />
+        ) : (
+          <Text style={styles.proceedButtonText}>Proceed</Text>
+        )}
+      </TouchableOpacity>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 20,
+    backgroundColor: 'white',
+    marginTop: 40.58,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginBottom: 20.45,
+  },
+  subtitle: {
+    fontSize: 13,
+    color: 'black',
+    textAlign: 'center',
+    marginBottom: 20,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ddd',
+    borderRadius: 8,
+    padding: 10,
+    marginBottom: 50,
+  },
+  proceedButton: {
+    backgroundColor: 'forestgreen',
+    padding: 15,
+    borderRadius: 71,
+    alignItems: 'center',
+  },
+  proceedButtonText: {
+    color: 'white',
+    fontWeight: 'bold',
+  },
+});
+
+export default checkEmail;
