+++ src/screens/updatePassword.js
@@ -1,0 +1,163 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, TouchableOpacity, StyleSheet, Alert, ActivityIndicator } from 'react-native';
+import { Feather } from '@expo/vector-icons';
+import axios from 'axios';  
+import { useNavigation } from '@react-navigation/native'
+
+import API_URL from '../config/api';
+import PasswordStrengthMeter from '../components/PasswordStrengthMeter'
+
+
+const updatePassword = ({ route }) => {
+  const navigation = useNavigation()
+
+  const [newPassword, setNewPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [showNewPassword, setShowNewPassword] = useState(false);
+  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
+  const [loading, setLoading] = useState(false);
+  const [isPasswordInputFocused, setIsPasswordInputFocused] = useState(false);
+
+  
+
+  const email = route.params?.email; // Get the email passed in params
+
+  const handleProceed = async () => {
+    if (newPassword !== confirmPassword) {
+      Alert.alert('Error', 'Passwords do not match.');
+      return;
+    }
+
+    if (newPassword.length < 6) {
+      Alert.alert('Error', 'Password must be at least 6 characters long.');
+      return;
+    }
+
+    setLoading(true);
+
+      try {
+      const response = await axios.post(`${API_URL}/update-password`, {
+        email,
+        newPassword,
+      }, {
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+
+      const data = response.data;
+
+      if (response.status === 200) {
+        Alert.alert('Success', 'Password updated successfully.');
+        console.log(response)
+        navigation.navigate ('SignIn')
+        // Optionally navigate to login or success screen
+      } else {
+        Alert.alert('Error', data.message || 'Something went wrong. Please try again.');
+      }
+    } catch (error) {
+      console.error(error);
+      Alert.alert('Error', 'Unable to connect to the server. Please try again.');
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Password Recovery</Text>
+      {/* Correct way to include the email */}
+      <Text style={styles.subtitle}>Create a new password for {email}</Text>
+
+      <View style={styles.passwordContainer}>
+        <TextInput
+          style={styles.passwordInput}
+          placeholder="New Password"
+          value={newPassword}
+          onChangeText={setNewPassword}
+          secureTextEntry={!showNewPassword}
+              onFocus={() => setIsPasswordInputFocused(true)}
+          onBlur={() => setIsPasswordInputFocused(false)}
+        />
+        <TouchableOpacity onPress={() => setShowNewPassword(!showNewPassword)} style={styles.eyeIcon}>
+          <Feather name={showNewPassword ? 'eye' : 'eye-off'} size={24} color="gray" />
+        </TouchableOpacity>
+      </View>
+
+      <View style={styles.passwordContainer}>
+        <TextInput
+          style={styles.passwordInput}
+          placeholder="Confirm Password"
+          value={confirmPassword}
+          onChangeText={setConfirmPassword}
+          secureTextEntry={!showConfirmPassword}
+        />
+        <TouchableOpacity onPress={() => setShowConfirmPassword(!showConfirmPassword)} style={styles.eyeIcon}>
+          <Feather name={showConfirmPassword ? 'eye' : 'eye-off'} size={24} color="gray" />
+        </TouchableOpacity>
+      </View>
+
+           <View>
+ {isPasswordInputFocused && (
+        <PasswordStrengthMeter password={newPassword} />
+      )}      </View>
+
+      <TouchableOpacity style={styles.proceedButton} onPress={handleProceed} disabled={loading}>
+        {loading ? (
+          <ActivityIndicator color="white" />
+        ) : (
+          <Text style={styles.proceedButtonText}>Proceed</Text>
+        )}
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 20,
+    backgroundColor: 'white',
+    marginTop: 40.58,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginBottom: 21,
+  },
+  subtitle: {
+    fontSize: 13,
+    color: 'black',
+    textAlign: 'center',
+    marginBottom: 20,
+  },
+  passwordContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderWidth: 1,
+    borderColor: '#ddd',
+    borderRadius: 8,
+    marginBottom: 20,
+  },
+  passwordInput: {
+    flex: 1,
+    padding: 10,
+  },
+  eyeIcon: {
+    padding: 10,
+  },
+  proceedButton: {
+    backgroundColor: 'forestgreen',
+    padding: 15,
+    borderRadius: 71,
+    alignItems: 'center',
+    marginTop: 30,
+  },
+  proceedButtonText: {
+    color: 'white',
+    fontWeight: 'bold',
+  },
+});
+
+export default updatePassword;
